+++ Libraries/LibJS/CMakeLists.txt
@@ -27,6 +27,7 @@ set(SOURCES
     Heap/Heap.cpp
     Heap/HeapBlock.cpp
     Heap/MarkedVector.cpp
+    Heap/WeakContainer.cpp
     Lexer.cpp
     MarkupGenerator.cpp
     Module.cpp
@@ -253,7 +254,6 @@ set(SOURCES
     Runtime/Utf16String.cpp
     Runtime/Value.cpp
     Runtime/VM.cpp
-    Runtime/WeakContainer.cpp
     Runtime/WeakMap.cpp
     Runtime/WeakMapConstructor.cpp
     Runtime/WeakMapPrototype.cpp
