+++ Libraries/LibJS/Runtime/FinalizationRegistry.h
@@ -8,12 +8,12 @@
 
 #include <AK/SinglyLinkedList.h>
 #include <LibJS/Heap/GCPtr.h>
+#include <LibJS/Heap/WeakContainer.h>
 #include <LibJS/Runtime/FunctionObject.h>
 #include <LibJS/Runtime/GlobalObject.h>
 #include <LibJS/Runtime/JobCallback.h>
 #include <LibJS/Runtime/Object.h>
 #include <LibJS/Runtime/Value.h>
-#include <LibJS/Runtime/WeakContainer.h>
 
 namespace JS {
