+++ Libraries/LibJS/Runtime/WeakMap.h
@@ -7,9 +7,9 @@
 #pragma once
 
 #include <AK/HashMap.h>
+#include <LibJS/Heap/WeakContainer.h>
 #include <LibJS/Runtime/GlobalObject.h>
 #include <LibJS/Runtime/Object.h>
-#include <LibJS/Runtime/WeakContainer.h>
 
 namespace JS {
